requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
pydantic==1.10.15
bcrypt==4.1.2
PyJWT==2.9.0
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.3.2
anyio==4.4.0
